LPS-117232 Turn DataEngineNativeObjectObserver into a inner class of DataEngineNativeObjectPortalExecutor
